requirements.txt
google-genai
googlemaps
flask
python-dotenv
pydantic
requests
